refactor(core): split writers and simplify report plumbing (no behavior change)
