+++ crates/floe-core/src/io/write/iceberg.rs
@@ -1,45 +1,33 @@
 use std::collections::HashMap;
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
+use std::path::Path;
 
-use arrow::array::{
-    ArrayRef, BooleanArray, Date32Array, Float32Array, Float64Array, Int32Array, Int64Array,
-    StringArray, Time64MicrosecondArray, TimestampMicrosecondArray,
-};
 use arrow::record_batch::RecordBatch;
-use aws_config::meta::region::RegionProviderChain;
-use aws_sdk_glue::config::Region as GlueRegion;
-use aws_sdk_glue::types::{
-    StorageDescriptor as GlueStorageDescriptor, TableInput as GlueTableInput,
-};
-use aws_sdk_glue::Client as GlueClient;
-use iceberg::arrow::{schema_to_arrow_schema, RecordBatchPartitionSplitter};
 use iceberg::memory::{MemoryCatalogBuilder, MEMORY_CATALOG_WAREHOUSE};
-use iceberg::spec::{
-    DataFileFormat, NestedField, PrimitiveType, Schema, Transform, Type, UnboundPartitionSpec,
-};
+use iceberg::spec::{Schema, UnboundPartitionSpec};
 use iceberg::transaction::{ApplyTransactionAction, Transaction};
-use iceberg::writer::base_writer::data_file_writer::DataFileWriterBuilder;
-use iceberg::writer::file_writer::location_generator::{
-    DefaultFileNameGenerator, DefaultLocationGenerator,
-};
-use iceberg::writer::file_writer::rolling_writer::RollingFileWriterBuilder;
-use iceberg::writer::file_writer::ParquetWriterBuilder;
-use iceberg::writer::partitioning::fanout_writer::FanoutWriter;
-use iceberg::writer::partitioning::PartitioningWriter;
-use iceberg::writer::{IcebergWriter, IcebergWriterBuilder};
-use iceberg::{Catalog, CatalogBuilder, NamespaceIdent, TableCreation, TableIdent};
-use polars::prelude::{DataFrame, DataType, Series, TimeUnit};
-use uuid::Uuid;
+use iceberg::{Catalog, CatalogBuilder, NamespaceIdent, TableIdent};
+use polars::prelude::DataFrame;
 
-use crate::checks::normalize;
 use crate::errors::RunError;
 use crate::io::format::{AcceptedSinkAdapter, AcceptedWriteMetrics, AcceptedWriteOutput};
-use crate::io::storage::{object_store::iceberg_store_config, ObjectRef, Target};
+use crate::io::storage::Target;
 use crate::{config, io, FloeResult};
 
 use super::metrics;
+
+mod context;
+mod data_files;
+mod glue;
+mod metadata;
+mod schema;
+
+use self::context::{
+    build_iceberg_write_context, create_table, ensure_namespace, sanitize_table_name,
+};
+use self::data_files::{iceberg_small_file_threshold_bytes, write_data_files};
+use self::glue::{load_glue_table_state, upsert_glue_table};
+use self::metadata::parse_metadata_version_from_location;
+use self::schema::{ensure_partition_spec_matches, ensure_schema_matches, prepare_iceberg_write};
 
 struct IcebergAcceptedAdapter;
 
@@ -348,826 +336,6 @@ async fn write_iceberg_table_async(
         iceberg_namespace: glue_catalog.as_ref().map(|cfg| cfg.namespace.clone()),
         iceberg_table: glue_catalog.as_ref().map(|cfg| cfg.table.clone()),
     })
-}
-
-fn iceberg_small_file_threshold_bytes(entity: &config::EntityConfig) -> u64 {
-    metrics::default_small_file_threshold_bytes(
-        entity
-            .sink
-            .accepted
-            .options
-            .as_ref()
-            .and_then(|options| options.max_size_per_file),
-    )
-}
-
-fn build_iceberg_write_context(
-    target: &Target,
-    entity: &config::EntityConfig,
-    mode: config::WriteMode,
-    mut remote: Option<&mut IcebergRemoteContext<'_>>,
-) -> FloeResult<IcebergWriteContext> {
-    if entity.sink.accepted.iceberg.is_some() && remote.is_none() {
-        return Err(Box::new(RunError(format!(
-            "iceberg catalog writes require runtime catalog context for entity {}",
-            entity.name
-        ))));
-    }
-    match target {
-        Target::Local { base_path, .. } => {
-            let table_root = PathBuf::from(base_path);
-            fs::create_dir_all(&table_root)?;
-            let metadata_location = latest_local_metadata_location(&table_root)?;
-            Ok(IcebergWriteContext {
-                table_root_uri: base_path.to_string(),
-                catalog_name: "floe_iceberg",
-                catalog_props: HashMap::new(),
-                metadata_location,
-                glue_catalog: None,
-            })
-        }
-        Target::S3 {
-            storage,
-            uri,
-            bucket,
-            base_key,
-        } => {
-            if let Some(ctx) = remote.as_mut() {
-                let ctx = &mut **ctx;
-                if let Some(glue_target) = ctx.catalogs.resolve_iceberg_target(
-                    ctx.resolver,
-                    entity,
-                    &entity.sink.accepted,
-                )? {
-                    if glue_target.catalog_type == "glue" {
-                        let catalog_target = Target::from_resolved(&glue_target.table_location)?;
-                        let store = iceberg_store_config(&catalog_target, ctx.resolver, entity)?;
-                        return Ok(IcebergWriteContext {
-                            table_root_uri: store.warehouse_location,
-                            catalog_name: "floe_iceberg",
-                            catalog_props: store.file_io_props,
-                            metadata_location: None,
-                            glue_catalog: Some(GlueIcebergCatalogConfig {
-                                catalog_name: glue_target.catalog_name,
-                                region: glue_target.region,
-                                database: glue_target.database,
-                                namespace: glue_target.namespace,
-                                table: glue_target.table,
-                            }),
-                        });
-                    }
-                }
-            }
-
-            let metadata_location = if matches!(mode, config::WriteMode::Append) {
-                match remote.as_mut() {
-                    Some(ctx) => {
-                        let ctx = &mut **ctx;
-                        let client = ctx.cloud.client_for(ctx.resolver, storage, entity)?;
-                        latest_s3_metadata_location(client, base_key)?
-                    }
-                    None => {
-                        let mut client = io::storage::s3::S3Client::new(bucket.clone(), None)?;
-                        latest_s3_metadata_location(&mut client, base_key)?
-                    }
-                }
-            } else {
-                None
-            };
-
-            match remote.as_mut() {
-                Some(ctx) => {
-                    let ctx = &mut **ctx;
-                    let store = iceberg_store_config(target, ctx.resolver, entity)?;
-                    Ok(IcebergWriteContext {
-                        table_root_uri: store.warehouse_location,
-                        catalog_name: "floe_iceberg",
-                        catalog_props: store.file_io_props,
-                        metadata_location,
-                        glue_catalog: None,
-                    })
-                }
-                None => Ok(IcebergWriteContext {
-                    table_root_uri: uri.clone(),
-                    catalog_name: "floe_iceberg",
-                    catalog_props: HashMap::new(),
-                    metadata_location,
-                    glue_catalog: None,
-                }),
-            }
-        }
-        Target::Gcs {
-            storage,
-            uri,
-            bucket,
-            base_key,
-        } => {
-            let metadata_location = if matches!(mode, config::WriteMode::Append) {
-                match remote.as_mut() {
-                    Some(ctx) => {
-                        let ctx = &mut **ctx;
-                        let client = ctx.cloud.client_for(ctx.resolver, storage, entity)?;
-                        latest_gcs_metadata_location(client, base_key)?
-                    }
-                    None => {
-                        let mut client = io::storage::gcs::GcsClient::new(bucket.clone())?;
-                        latest_gcs_metadata_location(&mut client, base_key)?
-                    }
-                }
-            } else {
-                None
-            };
-
-            match remote.as_mut() {
-                Some(ctx) => {
-                    let ctx = &mut **ctx;
-                    let store = iceberg_store_config(target, ctx.resolver, entity)?;
-                    Ok(IcebergWriteContext {
-                        table_root_uri: store.warehouse_location,
-                        catalog_name: "floe_iceberg",
-                        catalog_props: store.file_io_props,
-                        metadata_location,
-                        glue_catalog: None,
-                    })
-                }
-                None => Ok(IcebergWriteContext {
-                    table_root_uri: uri.clone(),
-                    catalog_name: "floe_iceberg",
-                    catalog_props: HashMap::new(),
-                    metadata_location,
-                    glue_catalog: None,
-                }),
-            }
-        }
-        Target::Adls { .. } => Err(Box::new(RunError(format!(
-            "iceberg sink currently supports local, s3, or gcs storage only for entity {}",
-            entity.name
-        )))),
-    }
-}
-
-async fn ensure_namespace(
-    catalog: &iceberg::MemoryCatalog,
-    namespace: &NamespaceIdent,
-) -> FloeResult<()> {
-    let exists = catalog
-        .namespace_exists(namespace)
-        .await
-        .map_err(map_iceberg_err("iceberg namespace exists check failed"))?;
-    if !exists {
-        catalog
-            .create_namespace(namespace, HashMap::new())
-            .await
-            .map_err(map_iceberg_err("iceberg namespace create failed"))?;
-    }
-    Ok(())
-}
-
-async fn create_table(
-    catalog: &iceberg::MemoryCatalog,
-    namespace: &NamespaceIdent,
-    table_ident: &TableIdent,
-    table_root: String,
-    schema: &Schema,
-    partition_spec: Option<UnboundPartitionSpec>,
-) -> FloeResult<iceberg::table::Table> {
-    let creation_builder = TableCreation::builder()
-        .name(table_ident.name().to_string())
-        .location(table_root)
-        .schema(schema.clone());
-    let creation = match partition_spec {
-        Some(partition_spec) => creation_builder.partition_spec(partition_spec).build(),
-        None => creation_builder.build(),
-    };
-    catalog
-        .create_table(namespace, creation)
-        .await
-        .map_err(map_iceberg_err("iceberg create table failed"))
-}
-
-async fn write_data_files(
-    table: &iceberg::table::Table,
-    batch: RecordBatch,
-) -> FloeResult<Vec<iceberg::spec::DataFile>> {
-    let location_generator = DefaultLocationGenerator::new(table.metadata().clone())
-        .map_err(map_iceberg_err("iceberg location generator failed"))?;
-    let file_name_generator = DefaultFileNameGenerator::new(
-        "floe".to_string(),
-        Some(Uuid::now_v7().to_string()),
-        DataFileFormat::Parquet,
-    );
-    let parquet_writer_builder = ParquetWriterBuilder::new(
-        Default::default(),
-        table.metadata().current_schema().clone(),
-    );
-    let rolling_writer_builder = RollingFileWriterBuilder::new_with_default_file_size(
-        parquet_writer_builder,
-        table.file_io().clone(),
-        location_generator,
-        file_name_generator,
-    );
-
-    let data_file_writer_builder = DataFileWriterBuilder::new(rolling_writer_builder);
-    let partition_spec = table.metadata().default_partition_spec().clone();
-    if partition_spec.is_unpartitioned() {
-        let mut writer = data_file_writer_builder
-            .build(None)
-            .await
-            .map_err(map_iceberg_err("iceberg data writer build failed"))?;
-        writer
-            .write(batch)
-            .await
-            .map_err(map_iceberg_err("iceberg data write failed"))?;
-        return writer
-            .close()
-            .await
-            .map_err(map_iceberg_err("iceberg data writer close failed"));
-    }
-
-    let splitter = RecordBatchPartitionSplitter::try_new_with_computed_values(
-        table.metadata().current_schema().clone(),
-        partition_spec,
-    )
-    .map_err(map_iceberg_err("iceberg partition splitter init failed"))?;
-    let partitioned_batches = splitter
-        .split(&batch)
-        .map_err(map_iceberg_err("iceberg partition split failed"))?;
-
-    let mut writer = FanoutWriter::new(data_file_writer_builder);
-    for (partition_key, partition_batch) in partitioned_batches {
-        writer
-            .write(partition_key, partition_batch)
-            .await
-            .map_err(map_iceberg_err("iceberg partitioned data write failed"))?;
-    }
-    writer.close().await.map_err(map_iceberg_err(
-        "iceberg partitioned data writer close failed",
-    ))
-}
-
-fn prepare_iceberg_write(
-    df: &DataFrame,
-    entity: &config::EntityConfig,
-) -> FloeResult<PreparedIcebergWrite> {
-    let columns = resolve_output_columns(df, entity)?;
-    if columns.is_empty() {
-        return Err(Box::new(RunError(format!(
-            "iceberg sink requires at least one column for entity {}",
-            entity.name
-        ))));
-    }
-
-    let mut field_id = 1_i32;
-    let mut iceberg_fields = Vec::with_capacity(columns.len());
-    let mut arrays = Vec::with_capacity(columns.len());
-
-    for (name, nullable, series) in columns {
-        if !nullable && series.null_count() > 0 {
-            return Err(Box::new(RunError(format!(
-                "iceberg write rejected nulls for non-nullable column {}",
-                name
-            ))));
-        }
-
-        let primitive = polars_dtype_to_iceberg_type(series, &entity.name)?;
-        let field = if nullable {
-            NestedField::optional(field_id, name.clone(), Type::Primitive(primitive))
-        } else {
-            NestedField::required(field_id, name.clone(), Type::Primitive(primitive))
-        };
-        iceberg_fields.push(field.into());
-        arrays.push(series_to_arrow_array(series, &name, &entity.name)?);
-        field_id += 1;
-    }
-
-    let iceberg_schema = Schema::builder()
-        .with_schema_id(1)
-        .with_fields(iceberg_fields)
-        .build()
-        .map_err(map_iceberg_err("iceberg schema build failed"))?;
-    let partition_spec = build_unbound_partition_spec(&iceberg_schema, entity)?;
-    let arrow_schema = Arc::new(
-        schema_to_arrow_schema(&iceberg_schema)
-            .map_err(map_iceberg_err("iceberg arrow schema conversion failed"))?,
-    );
-    let batch = RecordBatch::try_new(arrow_schema, arrays).map_err(|err| {
-        Box::new(RunError(format!(
-            "iceberg record batch build failed: {err}"
-        ))) as Box<dyn std::error::Error + Send + Sync>
-    })?;
-
-    Ok(PreparedIcebergWrite {
-        iceberg_schema,
-        partition_spec,
-        batch,
-    })
-}
-
-fn build_unbound_partition_spec(
-    iceberg_schema: &Schema,
-    entity: &config::EntityConfig,
-) -> FloeResult<Option<UnboundPartitionSpec>> {
-    let Some(fields) = entity.sink.accepted.partition_spec.as_ref() else {
-        return Ok(None);
-    };
-
-    let mut builder = UnboundPartitionSpec::builder();
-    for field in fields {
-        let column = field.column.trim();
-        let schema_field = iceberg_schema.field_by_name(column).ok_or_else(|| {
-            Box::new(RunError(format!(
-                "entity.name={} iceberg partition_spec column {} was not found in runtime schema",
-                entity.name, column
-            ))) as Box<dyn std::error::Error + Send + Sync>
-        })?;
-        let normalized_transform = field.transform.trim().to_ascii_lowercase();
-        let transform = iceberg_partition_transform(&normalized_transform, &entity.name, column)?;
-        let partition_field_name = iceberg_partition_field_name(column, &normalized_transform);
-        builder = builder
-            .add_partition_field(schema_field.id, partition_field_name, transform)
-            .map_err(map_iceberg_err("iceberg partition spec build failed"))?;
-    }
-
-    Ok(Some(builder.build()))
-}
-
-fn iceberg_partition_transform(
-    transform: &str,
-    entity_name: &str,
-    column: &str,
-) -> FloeResult<Transform> {
-    let iceberg_transform = match transform {
-        "identity" => Transform::Identity,
-        "year" => Transform::Year,
-        "month" => Transform::Month,
-        "day" => Transform::Day,
-        "hour" => Transform::Hour,
-        _ => {
-            return Err(Box::new(RunError(format!(
-            "entity.name={} iceberg partition_spec column {} has unsupported runtime transform {}",
-            entity_name, column, transform
-        ))))
-        }
-    };
-    Ok(iceberg_transform)
-}
-
-fn iceberg_partition_field_name(column: &str, transform: &str) -> String {
-    if transform == "identity" {
-        column.to_string()
-    } else {
-        format!("{column}_{transform}")
-    }
-}
-
-fn resolve_output_columns<'a>(
-    df: &'a DataFrame,
-    entity: &'a config::EntityConfig,
-) -> FloeResult<Vec<(String, bool, &'a Series)>> {
-    if entity.schema.columns.is_empty() {
-        let mut columns = Vec::with_capacity(df.width());
-        for column in df.get_columns() {
-            let series = column.as_materialized_series();
-            let name = series.name().to_string();
-            // Keep inferred schemas stable across runs when no explicit schema is provided.
-            // Batch-dependent null distributions should not flip required/optional.
-            let nullable = true;
-            columns.push((name, nullable, series));
-        }
-        return Ok(columns);
-    }
-
-    let schema_columns = normalize::resolve_output_columns(
-        &entity.schema.columns,
-        normalize::resolve_normalize_strategy(entity)?.as_deref(),
-    );
-
-    let mut columns = Vec::with_capacity(schema_columns.len());
-    for column in &schema_columns {
-        let series = df
-            .column(column.name.as_str())
-            .map_err(|err| Box::new(RunError(format!("iceberg column lookup failed: {err}"))))?
-            .as_materialized_series();
-        columns.push((column.name.clone(), column.nullable.unwrap_or(true), series));
-    }
-    Ok(columns)
-}
-
-fn polars_dtype_to_iceberg_type(series: &Series, entity_name: &str) -> FloeResult<PrimitiveType> {
-    let primitive = match series.dtype() {
-        DataType::String => PrimitiveType::String,
-        DataType::Boolean => PrimitiveType::Boolean,
-        DataType::Int8 | DataType::Int16 | DataType::Int32 => PrimitiveType::Int,
-        DataType::Int64 => PrimitiveType::Long,
-        DataType::Float32 => PrimitiveType::Float,
-        DataType::Float64 => PrimitiveType::Double,
-        DataType::Date => PrimitiveType::Date,
-        DataType::Time => PrimitiveType::Time,
-        DataType::Datetime(_, tz) => {
-            if tz.is_some() {
-                PrimitiveType::Timestamptz
-            } else {
-                PrimitiveType::Timestamp
-            }
-        }
-        dtype => {
-            return Err(Box::new(RunError(format!(
-                "iceberg sink supports scalar types only; unsupported dtype {dtype:?} for column {} (entity {})",
-                series.name(),
-                entity_name
-            ))))
-        }
-    };
-    Ok(primitive)
-}
-
-fn series_to_arrow_array(
-    series: &Series,
-    column_name: &str,
-    entity_name: &str,
-) -> FloeResult<ArrayRef> {
-    let array: ArrayRef = match series.dtype() {
-        DataType::String => Arc::new(StringArray::from_iter(series.str()?)),
-        DataType::Boolean => Arc::new(BooleanArray::from_iter(series.bool()?)),
-        DataType::Int8 => {
-            let values = series.i8()?;
-            Arc::new(Int32Array::from_iter(values.into_iter().map(|opt| opt.map(i32::from))))
-        }
-        DataType::Int16 => {
-            let values = series.i16()?;
-            Arc::new(Int32Array::from_iter(values.into_iter().map(|opt| opt.map(i32::from))))
-        }
-        DataType::Int32 => Arc::new(Int32Array::from_iter(series.i32()?)),
-        DataType::Int64 => Arc::new(Int64Array::from_iter(series.i64()?)),
-        DataType::Float32 => Arc::new(Float32Array::from_iter(series.f32()?)),
-        DataType::Float64 => Arc::new(Float64Array::from_iter(series.f64()?)),
-        DataType::Date => {
-            let values = series.date()?;
-            Arc::new(Date32Array::from_iter(values.phys.iter()))
-        }
-        DataType::Time => {
-            let values = series.time()?;
-            let micros = values.phys.iter().map(|opt| opt.map(|value| value / 1_000));
-            Arc::new(Time64MicrosecondArray::from_iter(micros))
-        }
-        DataType::Datetime(unit, _) => {
-            let values = series.datetime()?;
-            let micros = values.phys.iter().map(|opt| match unit {
-                TimeUnit::Milliseconds => opt.map(|value| value.saturating_mul(1_000)),
-                TimeUnit::Microseconds => opt,
-                TimeUnit::Nanoseconds => opt.map(|value| value / 1_000),
-            });
-            Arc::new(TimestampMicrosecondArray::from_iter(micros))
-        }
-        dtype => {
-            return Err(Box::new(RunError(format!(
-                "iceberg sink supports scalar types only; unsupported dtype {dtype:?} for column {} (entity {})",
-                column_name, entity_name
-            ))))
-        }
-    };
-    Ok(array)
-}
-
-fn ensure_schema_matches(
-    existing: &Schema,
-    expected: &Schema,
-    entity: &config::EntityConfig,
-) -> FloeResult<()> {
-    let existing_fields = existing.as_struct().fields();
-    let expected_fields = expected.as_struct().fields();
-    if existing_fields.len() != expected_fields.len() {
-        return Err(Box::new(RunError(format!(
-            "entity.name={} iceberg schema evolution is not supported (column count differs: existing={}, incoming={})",
-            entity.name,
-            existing_fields.len(),
-            expected_fields.len()
-        ))));
-    }
-
-    for (index, (existing_field, expected_field)) in existing_fields
-        .iter()
-        .zip(expected_fields.iter())
-        .enumerate()
-    {
-        if existing_field.name != expected_field.name
-            || existing_field.required != expected_field.required
-            || existing_field.field_type != expected_field.field_type
-        {
-            return Err(Box::new(RunError(format!(
-                "entity.name={} iceberg schema evolution is not supported (column {} differs: existing={} incoming={})",
-                entity.name,
-                index,
-                describe_field(existing_field),
-                describe_field(expected_field)
-            ))));
-        }
-    }
-
-    Ok(())
-}
-
-fn ensure_partition_spec_matches(
-    existing: &iceberg::spec::PartitionSpec,
-    expected: Option<&UnboundPartitionSpec>,
-    expected_schema: &Schema,
-    entity: &config::EntityConfig,
-) -> FloeResult<()> {
-    let Some(expected_unbound) = expected else {
-        return Ok(());
-    };
-
-    let expected_bound = expected_unbound
-        .clone()
-        .bind(Arc::new(expected_schema.clone()))
-        .map_err(map_iceberg_err("iceberg partition spec bind failed"))?;
-
-    if !existing.is_compatible_with(&expected_bound) {
-        return Err(Box::new(RunError(format!(
-            "entity.name={} iceberg partition spec evolution is not supported (existing={} incoming={})",
-            entity.name,
-            describe_partition_spec(existing),
-            describe_partition_spec(&expected_bound)
-        ))));
-    }
-
-    Ok(())
-}
-
-fn describe_field(field: &Arc<iceberg::spec::NestedField>) -> String {
-    let required = if field.required {
-        "required"
-    } else {
-        "optional"
-    };
-    format!("{}:{}:{required}", field.name, field.field_type)
-}
-
-fn describe_partition_spec(spec: &iceberg::spec::PartitionSpec) -> String {
-    if spec.fields().is_empty() {
-        return "unpartitioned".to_string();
-    }
-
-    spec.fields()
-        .iter()
-        .map(|field| format!("{}:{}:{:?}", field.source_id, field.name, field.transform))
-        .collect::<Vec<_>>()
-        .join(",")
-}
-
-fn latest_local_metadata_location(table_root: &Path) -> FloeResult<Option<String>> {
-    let metadata_dir = table_root.join("metadata");
-    if !metadata_dir.exists() {
-        return Ok(None);
-    }
-
-    let mut best: Option<(i64, PathBuf)> = None;
-    for entry in fs::read_dir(&metadata_dir)? {
-        let entry = entry?;
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let Some(file_name) = path.file_name().and_then(|name| name.to_str()) else {
-            continue;
-        };
-        if !file_name.ends_with(".metadata.json") {
-            continue;
-        }
-        let Some(version) = parse_metadata_version_from_filename(file_name) else {
-            continue;
-        };
-        let replace = match &best {
-            None => true,
-            Some((best_version, best_path)) => {
-                version > *best_version || (version == *best_version && path > *best_path)
-            }
-        };
-        if replace {
-            best = Some((version, path));
-        }
-    }
-
-    Ok(best.map(|(_, path)| path.display().to_string()))
-}
-
-fn latest_s3_metadata_location(
-    client: &mut dyn io::storage::StorageClient,
-    base_key: &str,
-) -> FloeResult<Option<String>> {
-    let metadata_prefix = if base_key.trim_matches('/').is_empty() {
-        "metadata/".to_string()
-    } else {
-        format!("{}/metadata/", base_key.trim_matches('/'))
-    };
-    let listed = client.list(&metadata_prefix)?;
-    latest_metadata_location_from_objects(listed)
-}
-
-fn latest_gcs_metadata_location(
-    client: &mut dyn io::storage::StorageClient,
-    base_key: &str,
-) -> FloeResult<Option<String>> {
-    let metadata_prefix = if base_key.trim_matches('/').is_empty() {
-        "metadata/".to_string()
-    } else {
-        format!("{}/metadata/", base_key.trim_matches('/'))
-    };
-    let listed = client.list(&metadata_prefix)?;
-    latest_metadata_location_from_objects(listed)
-}
-
-fn latest_metadata_location_from_objects(objects: Vec<ObjectRef>) -> FloeResult<Option<String>> {
-    let mut best: Option<(i64, String, String)> = None;
-    for object in objects {
-        let file_name = object
-            .key
-            .rsplit('/')
-            .next()
-            .unwrap_or(object.key.as_str())
-            .to_string();
-        if !file_name.ends_with(".metadata.json") {
-            continue;
-        }
-        let Some(version) = parse_metadata_version_from_filename(&file_name) else {
-            continue;
-        };
-        let replace = match &best {
-            None => true,
-            Some((best_version, best_key, _)) => {
-                version > *best_version || (version == *best_version && object.key > *best_key)
-            }
-        };
-        if replace {
-            best = Some((version, object.key.clone(), object.uri.clone()));
-        }
-    }
-    Ok(best.map(|(_, _, uri)| uri))
-}
-
-#[derive(Debug, Clone)]
-struct GlueTableState {
-    metadata_location: Option<String>,
-    version_id: Option<String>,
-}
-
-async fn build_glue_client(region: &str) -> FloeResult<GlueClient> {
-    let region_provider =
-        RegionProviderChain::first_try(GlueRegion::new(region.to_string())).or_default_provider();
-    let config = aws_config::defaults(aws_config::BehaviorVersion::latest())
-        .region(region_provider)
-        .load()
-        .await;
-    Ok(GlueClient::new(&config))
-}
-
-async fn load_glue_table_state(glue_cfg: &GlueIcebergCatalogConfig) -> FloeResult<GlueTableState> {
-    let client = build_glue_client(&glue_cfg.region).await?;
-    let response = client
-        .get_table()
-        .database_name(glue_cfg.database.as_str())
-        .name(glue_cfg.table.as_str())
-        .send()
-        .await;
-    match response {
-        Ok(output) => {
-            let table = output.table().ok_or_else(|| {
-                Box::new(RunError(format!(
-                    "glue get_table returned no table for {}.{}",
-                    glue_cfg.database, glue_cfg.table
-                ))) as Box<dyn std::error::Error + Send + Sync>
-            })?;
-            let parameters = table.parameters();
-            let metadata_location =
-                parameters.and_then(|params| params.get("metadata_location").cloned());
-            let iceberg_param = parameters
-                .and_then(|params| params.get("table_type"))
-                .map(|value| value.eq_ignore_ascii_case("ICEBERG"))
-                .unwrap_or(false);
-            if !iceberg_param || metadata_location.is_none() {
-                return Err(Box::new(RunError(format!(
-                    "glue table {}.{} exists but is not an Iceberg table managed by Floe (missing Iceberg parameters/metadata_location)",
-                    glue_cfg.database, glue_cfg.table
-                ))));
-            }
-            Ok(GlueTableState {
-                metadata_location,
-                version_id: table.version_id().map(ToOwned::to_owned),
-            })
-        }
-        Err(err) => {
-            if err
-                .as_service_error()
-                .is_some_and(|service_err| service_err.is_entity_not_found_exception())
-            {
-                return Ok(GlueTableState {
-                    metadata_location: None,
-                    version_id: None,
-                });
-            }
-            Err(Box::new(RunError(format!(
-                "glue get_table failed for {}.{}: {err}",
-                glue_cfg.database, glue_cfg.table
-            ))))
-        }
-    }
-}
-
-async fn upsert_glue_table(
-    glue_cfg: &GlueIcebergCatalogConfig,
-    table_root_uri: &str,
-    metadata_location: &str,
-    version_id: Option<&str>,
-) -> FloeResult<()> {
-    let client = build_glue_client(&glue_cfg.region).await?;
-    let table_input = build_glue_table_input(glue_cfg, table_root_uri, metadata_location);
-
-    let create_result = client
-        .create_table()
-        .database_name(glue_cfg.database.as_str())
-        .name(glue_cfg.table.as_str())
-        .table_input(table_input.clone())
-        .send()
-        .await;
-    match create_result {
-        Ok(_) => return Ok(()),
-        Err(err) => {
-            if !err
-                .as_service_error()
-                .is_some_and(|service_err| service_err.is_already_exists_exception())
-            {
-                return Err(Box::new(RunError(format!(
-                    "glue create_table failed for {}.{}: {err}",
-                    glue_cfg.database, glue_cfg.table
-                ))));
-            }
-        }
-    }
-
-    let mut update = client
-        .update_table()
-        .database_name(glue_cfg.database.as_str())
-        .name(glue_cfg.table.as_str())
-        .table_input(table_input)
-        .skip_archive(true);
-    if let Some(version_id) = version_id {
-        update = update.version_id(version_id);
-    }
-    update.send().await.map_err(|err| {
-        Box::new(RunError(format!(
-            "glue update_table failed for {}.{}: {err}",
-            glue_cfg.database, glue_cfg.table
-        ))) as Box<dyn std::error::Error + Send + Sync>
-    })?;
-    Ok(())
-}
-
-fn build_glue_table_input(
-    glue_cfg: &GlueIcebergCatalogConfig,
-    table_root_uri: &str,
-    metadata_location: &str,
-) -> GlueTableInput {
-    let storage_descriptor = GlueStorageDescriptor::builder()
-        .location(table_root_uri)
-        .build();
-
-    GlueTableInput::builder()
-        .name(glue_cfg.table.as_str())
-        .table_type("EXTERNAL_TABLE")
-        .storage_descriptor(storage_descriptor)
-        .set_partition_keys(Some(Vec::new()))
-        .parameters("table_type", "ICEBERG")
-        .parameters("EXTERNAL", "TRUE")
-        .parameters("metadata_location", metadata_location)
-        .parameters("floe.iceberg.namespace", glue_cfg.namespace.as_str())
-        .build()
-        .expect("glue table input builder validated")
-}
-
-fn parse_metadata_version_from_location(location: &str) -> Option<i64> {
-    let file_name = Path::new(location).file_name()?.to_str()?;
-    parse_metadata_version_from_filename(file_name)
-}
-
-fn parse_metadata_version_from_filename(file_name: &str) -> Option<i64> {
-    let prefix = file_name.split_once('-')?.0;
-    prefix.parse::<i64>().ok()
-}
-
-fn sanitize_table_name(name: &str) -> String {
-    let mut out = String::with_capacity(name.len());
-    for ch in name.chars() {
-        if ch.is_ascii_alphanumeric() || ch == '_' || ch == '-' {
-            out.push(ch);
-        } else {
-            out.push('_');
-        }
-    }
-    if out.is_empty() {
-        "table".to_string()
-    } else {
-        out
-    }
 }
 
 fn map_iceberg_err(
